move everything to a class
